backend/requirements.txt
fastapi
uvicorn
pymongo
motor
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
supabase
email-validator
bcrypt==4.0.1
pillow
aiofiles
onnxruntime
opencv-python
numpy
scipy
torch
torchvision
scikit-image
tqdm
rembg
ultralytics
